requirements_heroku.txt
aiofiles==0.6.0
aioflask==0.1.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.3.0
awswrangler==2.3.0
beautifulsoup4==4.9.1
blinker==1.4
boto3==1.16.57
botocore==1.19.57
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
et-xmlfile==1.0.1
fastapi==0.63.0
Flask==1.1.2
greenlet==1.0.0
greenletio==0.0.7
h11==0.12.0
h2==4.0.0
hpack==4.0.0
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.5.2
MarkupSafe==1.1.1
multidict==4.7.6
nest-asyncio==1.4.3
numexpr==2.7.2
numpy==1.19.5
openpyxl==3.0.6
pandas
pg8000==1.16.6
priority==1.3.0
pyarrow==2.0.0
pydantic==1.7.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.5
redshift-connector==2.0.872
requests==2.25.1
s3transfer==0.3.4
scramp==1.2.0
search-engine-parser==0.6.2
six==1.15.0
smart-open==4.1.2
soupsieve==2.1
starlette==0.13.6
tables==3.6.1
toml==0.10.2
urllib3==1.26.2
uvicorn==0.13.3
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.6.3
sqlalchemy
psycopg2
fastapi-login
scraperapi-sdk
python-jose[cryptography]